(android) Fix modal bottom sheet issues following compose update
