+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/components/Card.kt
@@ -50,6 +50,7 @@ import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import fr.acinq.phoenix.android.components.dialogs.IconPopup
 
 @Composable
 fun Card(
