+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/components/dialogs/BasicDialog.kt
@@ -1,0 +1,97 @@
+/*
+ * Copyright 2023 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.acinq.phoenix.android.components.dialogs
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ColumnScope
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.widthIn
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.DialogProperties
+import fr.acinq.phoenix.android.R
+import fr.acinq.phoenix.android.components.Button
+
+@Composable
+fun Dialog(
+    onDismiss: () -> Unit,
+    title: String? = null,
+    properties: DialogProperties = DialogProperties(usePlatformDefaultWidth = false),
+    isScrollable: Boolean = true,
+    buttonsTopMargin: Dp = 20.dp,
+    buttons: (@Composable RowScope.() -> Unit)? = { Button(onClick = onDismiss, text = stringResource(id = R.string.btn_ok), padding = PaddingValues(16.dp), shape = RoundedCornerShape(16.dp)) },
+    content: @Composable ColumnScope.() -> Unit,
+) {
+    androidx.compose.ui.window.Dialog(onDismissRequest = onDismiss, properties = properties) {
+        DialogBody(isScrollable) {
+            // optional title
+            title?.run {
+                Text(text = title, modifier = Modifier.padding(start = 24.dp, end = 24.dp, top = 20.dp, bottom = 12.dp), style = MaterialTheme.typography.h4)
+            }
+            // content, must set the padding etc...
+            content()
+            // buttons
+            if (buttons != null) {
+                Spacer(Modifier.height(buttonsTopMargin))
+                Row(modifier = Modifier
+                    .align(Alignment.End)
+                    .padding(8.dp)) {
+                    buttons()
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun DialogBody(
+    isScrollable: Boolean = true,
+    content: @Composable ColumnScope.() -> Unit,
+) {
+    Column(
+        Modifier
+            .padding(vertical = 16.dp, horizontal = 16.dp) // min padding for tall/wide dialogs
+            .clip(MaterialTheme.shapes.large)
+            .background(MaterialTheme.colors.surface)
+            .widthIn(max = 600.dp)
+            .then(
+                if (isScrollable) {
+                    Modifier.verticalScroll(rememberScrollState())
+                } else {
+                    Modifier
+                }
+            )
+    ) {
+        content()
+    }
+}
