+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/home/releasenotes/ReleaseNotesDialog.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import fr.acinq.phoenix.android.R
-import fr.acinq.phoenix.android.components.Dialog
+import fr.acinq.phoenix.android.components.dialogs.Dialog
 import fr.acinq.phoenix.android.internalData
 import kotlinx.coroutines.launch
