+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/payments/receive/ReceiveOfferView.kt
@@ -19,25 +19,20 @@ package fr.acinq.phoenix.android.payments.receive
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.selection.SelectionContainer
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -58,75 +53,61 @@ import fr.acinq.phoenix.android.components.BorderButton
 import fr.acinq.phoenix.android.components.Clickable
 import fr.acinq.phoenix.android.components.PhoenixIcon
 import fr.acinq.phoenix.android.components.TextWithIcon
+import fr.acinq.phoenix.android.components.dialogs.ModalBottomSheet
 import fr.acinq.phoenix.android.components.openLink
 import fr.acinq.phoenix.android.internalData
 import fr.acinq.phoenix.android.utils.copyToClipboard
 import fr.acinq.phoenix.android.utils.share
 
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun DisplayOfferDialog(
     onDismiss: () -> Unit,
     offerState: OfferState,
 ) {
     val context = LocalContext.current
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
 
     val bip353AddressState = internalData.getBip353Address.collectAsState(initial = null)
     val address = bip353AddressState.value
 
     ModalBottomSheet(
-        sheetState = sheetState,
-        onDismissRequest = {
-            // executed when user click outside the sheet, and after sheet has been hidden thru state.
-            onDismiss()
-        },
-        modifier = Modifier.heightIn(max = 700.dp),
-        containerColor = MaterialTheme.colors.surface,
-        contentColor = MaterialTheme.colors.onSurface,
-        scrimColor = MaterialTheme.colors.onBackground.copy(alpha = 0.2f),
+        onDismiss = onDismiss,
+        skipPartiallyExpanded = true,
+        horizontalAlignment = Alignment.CenterHorizontally,
+        internalPadding = PaddingValues(top = 0.dp, start = 12.dp, end = 12.dp, bottom = 50.dp),
     ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier
-                .fillMaxWidth()
-                .verticalScroll(rememberScrollState())
-                .padding(top = 0.dp, start = 12.dp, end = 12.dp, bottom = 50.dp)
-        ) {
-            when (offerState) {
-                OfferState.Init -> {
-                    Text(text = stringResource(id = R.string.utils_loading_data))
+        when (offerState) {
+            OfferState.Init -> {
+                Text(text = stringResource(id = R.string.utils_loading_data))
+            }
+            is OfferState.Show -> {
+                HowToOffer()
+                Spacer(modifier = Modifier.height(6.dp))
+                Box(modifier = Modifier
+                    .widthIn(max = 350.dp)
+                    .clip(shape = RoundedCornerShape(16.dp))) {
+                    QRCodeImage(bitmap = offerState.bitmap, onLongClick = { copyToClipboard(context, offerState.encoded) })
                 }
-                is OfferState.Show -> {
-                    HowToOffer()
-                    Spacer(modifier = Modifier.height(6.dp))
-                    Box(modifier = Modifier
-                        .widthIn(max = 350.dp)
-                        .clip(shape = RoundedCornerShape(16.dp))) {
-                        QRCodeImage(bitmap = offerState.bitmap, onLongClick = { copyToClipboard(context, offerState.encoded) })
+                Bip353AddressDisplay(address)
+                Row {
+                    var showCopyDialog by remember { mutableStateOf(false) }
+                    if (showCopyDialog) {
+                        CopyAddressDialog(address = address, offer = offerState.encoded, onDismiss = { showCopyDialog = false })
                     }
-                    Bip353AddressDisplay(address)
-                    Row {
-                        var showCopyDialog by remember { mutableStateOf(false) }
-                        if (showCopyDialog && !address.isNullOrBlank()) {
-                            CopyAddressDialog(address = address, offer = offerState.encoded, onDismiss = { showCopyDialog = false })
-                        }
-                        BorderButton(
-                            text = stringResource(id = R.string.btn_copy),
-                            icon = R.drawable.ic_copy,
-                            onClick = { if (!address.isNullOrBlank()) showCopyDialog = true else copyToClipboard(context, data = offerState.encoded) },
-                        )
-                        Spacer(modifier = Modifier.width(16.dp))
-                        BorderButton(
-                            text = stringResource(id = R.string.btn_share),
-                            icon = R.drawable.ic_share,
-                            onClick = { share(context, "bitcoin:?lno=${offerState.encoded}", context.getString(R.string.receive_offer_share_subject), context.getString(R.string.receive_offer_share_title)) },
-                        )
-                    }
-                    Spacer(modifier = Modifier.height(16.dp))
-                    TorWarning()
+                    BorderButton(
+                        text = stringResource(id = R.string.btn_copy),
+                        icon = R.drawable.ic_copy,
+                        onClick = { showCopyDialog = true },
+                    )
+                    Spacer(modifier = Modifier.width(16.dp))
+                    BorderButton(
+                        text = stringResource(id = R.string.btn_share),
+                        icon = R.drawable.ic_share,
+                        onClick = { share(context, "bitcoin:?lno=${offerState.encoded}", context.getString(R.string.receive_offer_share_subject), context.getString(R.string.receive_offer_share_title)) },
+                    )
                 }
+                Spacer(modifier = Modifier.height(16.dp))
+                TorWarning()
             }
         }
     }
@@ -163,31 +144,21 @@ fun Bip353AddressDisplay(address: String?) {
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 private fun CopyAddressDialog(
-    address: String,
+    address: String?,
     offer: String,
     onDismiss: () -> Unit
 ) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-
     ModalBottomSheet(
-        sheetState = sheetState,
-        onDismissRequest = {
-            // executed when user click outside the sheet, and after sheet has been hidden thru state.
-            onDismiss()
-        },
-        modifier = Modifier.heightIn(max = 700.dp),
-        containerColor = MaterialTheme.colors.surface,
-        contentColor = MaterialTheme.colors.onSurface,
-        scrimColor = MaterialTheme.colors.onBackground.copy(alpha = 0.2f),
+        onDismiss = onDismiss,
+        internalPadding = PaddingValues(bottom = 50.dp),
     ) {
-        Column(Modifier.padding(bottom = 50.dp)) {
+        if (!address.isNullOrBlank()) {
             CopyButtonDialog(label = stringResource(id = R.string.receive_offer_copy_bip353), valueToCopy = stringResource(id = R.string.utils_bip353_with_prefix, address))
-            CopyButtonDialog(label = stringResource(id = R.string.receive_offer_copy_bolt12), valueToCopy = offer)
-            CopyButtonDialog(label = stringResource(id = R.string.receive_offer_copy_bip21), valueToCopy = "bitcoin:?lno=$offer")
         }
+        CopyButtonDialog(label = stringResource(id = R.string.receive_offer_copy_bolt12), valueToCopy = offer)
+        CopyButtonDialog(label = stringResource(id = R.string.receive_offer_copy_bip21), valueToCopy = "bitcoin:?lno=$offer")
     }
 }
 
@@ -208,9 +179,9 @@ private fun CopyButtonDialog(label: String, valueToCopy: String) {
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun HowToOffer() {
+    val context = LocalContext.current
     var showHowToDialog by remember { mutableStateOf(false) }
 
     Clickable(onClick = { showHowToDialog = true }, shape = RoundedCornerShape(12.dp)) {
@@ -221,48 +192,33 @@ fun HowToOffer() {
         }
     }
 
-    val context = LocalContext.current
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-
     if (showHowToDialog) {
         ModalBottomSheet(
-            sheetState = sheetState,
-            onDismissRequest = {
-                // executed when user click outside the sheet, and after sheet has been hidden thru state.
-                showHowToDialog = false
-            },
-            modifier = Modifier.heightIn(max = 700.dp),
-            containerColor = MaterialTheme.colors.surface,
-            contentColor = MaterialTheme.colors.onSurface,
-            scrimColor = MaterialTheme.colors.onBackground.copy(alpha = 0.2f),
+            onDismiss = { showHowToDialog = false },
+            skipPartiallyExpanded = true,
+            internalPadding = PaddingValues(top = 0.dp, start = 24.dp, end = 24.dp, bottom = 50.dp)
         ) {
-            Column(
-                modifier = Modifier
-                    .verticalScroll(rememberScrollState())
-                    .padding(top = 0.dp, start = 24.dp, end = 24.dp, bottom = 50.dp),
-            ) {
-                Text(text = stringResource(id = R.string.receive_offer_howto_details_1), style = MaterialTheme.typography.body2)
-                Spacer(modifier = Modifier.height(12.dp))
-                Text(text = stringResource(id = R.string.receive_offer_howto_details_2))
-                Spacer(modifier = Modifier.height(12.dp))
-                Text(text = stringResource(id = R.string.receive_offer_howto_details_3))
-                Spacer(modifier = Modifier.height(16.dp))
-                Text(text = stringResource(id = R.string.receive_offer_howto_details_4_title), style = MaterialTheme.typography.body2)
-                Spacer(modifier = Modifier.height(8.dp))
-                Text(text = stringResource(id = R.string.receive_offer_howto_details_4))
-                Spacer(modifier = Modifier.height(16.dp))
-                Text(text = stringResource(id = R.string.receive_offer_howto_details_5_title), style = MaterialTheme.typography.body2)
-                Spacer(modifier = Modifier.height(8.dp))
-                Text(text = stringResource(id = R.string.receive_offer_howto_details_5))
-                Spacer(modifier = Modifier.height(24.dp))
-                BorderButton(
-                    text = stringResource(id = R.string.receive_offer_howto_details_faq_link),
-                    icon = R.drawable.ic_external_link,
-                    onClick = { openLink(context, "https://bolt12.org") },
-                    shape = CircleShape,
-                    modifier = Modifier.align(Alignment.CenterHorizontally)
-                )
-            }
+            Text(text = stringResource(id = R.string.receive_offer_howto_details_1), style = MaterialTheme.typography.body2)
+            Spacer(modifier = Modifier.height(12.dp))
+            Text(text = stringResource(id = R.string.receive_offer_howto_details_2))
+            Spacer(modifier = Modifier.height(12.dp))
+            Text(text = stringResource(id = R.string.receive_offer_howto_details_3))
+            Spacer(modifier = Modifier.height(16.dp))
+            Text(text = stringResource(id = R.string.receive_offer_howto_details_4_title), style = MaterialTheme.typography.body2)
+            Spacer(modifier = Modifier.height(8.dp))
+            Text(text = stringResource(id = R.string.receive_offer_howto_details_4))
+            Spacer(modifier = Modifier.height(16.dp))
+            Text(text = stringResource(id = R.string.receive_offer_howto_details_5_title), style = MaterialTheme.typography.body2)
+            Spacer(modifier = Modifier.height(8.dp))
+            Text(text = stringResource(id = R.string.receive_offer_howto_details_5))
+            Spacer(modifier = Modifier.height(24.dp))
+            BorderButton(
+                text = stringResource(id = R.string.receive_offer_howto_details_faq_link),
+                icon = R.drawable.ic_external_link,
+                onClick = { openLink(context, "https://bolt12.org") },
+                shape = CircleShape,
+                modifier = Modifier.align(Alignment.CenterHorizontally)
+            )
         }
     }
 }
