+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/payments/send/spliceout/SpliceOutView.kt
@@ -26,9 +26,6 @@ import androidx.compose.foundation.text.selection.SelectionContainer
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -49,6 +46,7 @@ import fr.acinq.phoenix.android.LocalBitcoinUnit
 import fr.acinq.phoenix.android.R
 import fr.acinq.phoenix.android.business
 import fr.acinq.phoenix.android.components.*
+import fr.acinq.phoenix.android.components.dialogs.ModalBottomSheet
 import fr.acinq.phoenix.android.components.feedback.ErrorMessage
 import fr.acinq.phoenix.android.internalData
 import fr.acinq.phoenix.android.utils.Converter.toPrettyString
@@ -335,7 +333,6 @@ fun ColumnScope.SpliceOutCapacityDisclaimer(onConfirm: () -> Unit, onCancel: () 
     )
 }
 
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
 @Composable
 private fun ReviewSpliceOutAndConfirm(
     onExecute: () -> Unit,
@@ -345,20 +342,13 @@ private fun ReviewSpliceOutAndConfirm(
     showSpliceoutCapacityDisclaimer: Boolean,
 ) {
     val scope = rememberCoroutineScope()
-    val sheetState = rememberModalBottomSheetState()
     var showSheet by remember { mutableStateOf(false) }
 
     if (showSheet) {
         ModalBottomSheet(
-            sheetState = sheetState,
-            onDismissRequest = {
-                // executed when user click outside the sheet, and after sheet has been hidden thru state.
-                showSheet = false
-            },
-            modifier = Modifier.heightIn(max = 700.dp),
-            containerColor = MaterialTheme.colors.surface,
-            contentColor = MaterialTheme.colors.onSurface,
-            scrimColor = MaterialTheme.colors.onBackground.copy(alpha = 0.1f),
+            onDismiss = { showSheet = false },
+            internalPadding = PaddingValues(0.dp),
+            isContentScrollable = false
         ) {
             val pagerState = rememberPagerState(
                 initialPage = when {
