+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/settings/MutualCloseView.kt
@@ -36,6 +36,7 @@ import fr.acinq.lightning.utils.msat
 import fr.acinq.phoenix.android.*
 import fr.acinq.phoenix.android.R
 import fr.acinq.phoenix.android.components.*
+import fr.acinq.phoenix.android.components.dialogs.ConfirmDialog
 import fr.acinq.phoenix.android.components.mvi.MVIView
 import fr.acinq.phoenix.android.components.scanner.ScannerView
 import fr.acinq.phoenix.android.utils.Converter.toPrettyString
