+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/settings/channels/ImportChannelsData.kt
@@ -39,6 +39,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import fr.acinq.phoenix.android.R
 import fr.acinq.phoenix.android.business
 import fr.acinq.phoenix.android.components.*
+import fr.acinq.phoenix.android.components.dialogs.Dialog
 import fr.acinq.phoenix.android.components.feedback.ErrorMessage
 import fr.acinq.phoenix.android.utils.positiveColor
 import fr.acinq.phoenix.utils.channels.ChannelsImportResult
