+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/settings/electrum/ElectrumView.kt
@@ -41,6 +41,7 @@ import fr.acinq.lightning.utils.ServerAddress
 import fr.acinq.phoenix.android.*
 import fr.acinq.phoenix.android.R
 import fr.acinq.phoenix.android.components.*
+import fr.acinq.phoenix.android.components.dialogs.Dialog
 import fr.acinq.phoenix.android.components.feedback.ErrorMessage
 import fr.acinq.phoenix.android.components.mvi.MVIView
 import fr.acinq.phoenix.android.components.settings.Setting
